Move to RSpec::Parameterized::Core
