+++ lib/rspec-parameterized-core.rb
@@ -1,0 +1,1 @@
+require 'rspec/parameterized/core'
