--- lib/rspec-parameterized.rb
@@ -1,1 +1,0 @@
-require 'rspec/parameterized'
