--- lib/rspec/parameterized/arg.rb
@@ -1,7 +1,0 @@
-module RSpec
-  module Parameterized
-    class Arg
-      def apply(obj) ; end
-    end
-  end
-end
