+++ lib/rspec/parameterized/core/arg.rb
@@ -1,0 +1,9 @@
+module RSpec
+  module Parameterized
+    module Core
+      class Arg
+        def apply(obj) ; end
+      end
+    end
+  end
+end
