+++ lib/rspec/parameterized/core/ref_arg.rb
@@ -1,0 +1,20 @@
+require 'rspec/parameterized/core/arg'
+module RSpec
+  module Parameterized
+    module Core
+      class RefArg < Arg
+        def initialize(symbol)
+          @symbol = symbol
+        end
+
+        def apply(obj)
+          obj.send(@symbol)
+        end
+
+        def inspect
+          "#{@symbol}"
+        end
+      end
+    end
+  end
+end
